Update Window and Renderer

+++ EngineAPI/src/main/java/unknowndomain/engine/client/rendering/Renderer.java
@@ -10,5 +10,5 @@ public interface Renderer extends Disposable {
 
     void init(RenderContext context);
 
-    void render(double partial);
+    void render(float partial);
 }
